@@ -1,6 +1,6 @@
 {
   "name": "eslint-visitor-keys",
-  "version": "3.3.0",
+  "version": "3.4.1",
   "description": "Constants and utilities about visitor keys to traverse AST.",
   "type": "module",
   "main": "dist/eslint-visitor-keys.cjs",
@@ -19,30 +19,41 @@
     "dist/index.d.ts",
     "dist/visitor-keys.d.ts",
     "dist/eslint-visitor-keys.cjs",
+    "dist/eslint-visitor-keys.d.cts",
     "lib"
   ],
   "engines": {
     "node": "^12.22.0 || ^14.17.0 || >=16.0.0"
   },
   "devDependencies": {
-    "c8": "^7.7.3",
+    "@types/estree": "^0.0.51",
+    "@types/estree-jsx": "^0.0.1",
+    "@typescript-eslint/parser": "^5.14.0",
+    "c8": "^7.11.0",
+    "chai": "^4.3.6",
     "eslint": "^7.29.0",
     "eslint-config-eslint": "^7.0.0",
     "eslint-plugin-jsdoc": "^35.4.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-release": "^3.2.0",
-    "mocha": "^9.0.1",
+    "esquery": "^1.4.0",
+    "json-diff": "^0.7.3",
+    "mocha": "^9.2.1",
     "opener": "^1.5.2",
-    "rollup": "^2.52.1",
+    "rollup": "^2.70.0",
+    "rollup-plugin-dts": "^4.2.3",
     "tsd": "^0.19.1",
-    "typescript": "^4.5.5"
+    "typescript": "^4.6.2"
   },
   "scripts": {
     "prepare": "npm run build",
-    "build": "rollup -c && npm run tsc",
+    "bundle": "rollup -c",
+    "build": "npm run bundle && npm run tsc",
+    "build:debug": "npm run bundle -- -m && npm run tsc",
     "lint": "eslint .",
     "tsc": "tsc",
     "tsd": "tsd",
+    "build-keys": "node tools/build-keys-from-ts",
     "test": "mocha tests/lib/**/*.cjs && c8 mocha tests/lib/**/*.js && npm run tsd",
     "coverage": "c8 report --reporter lcov && opener coverage/lcov-report/index.html",
     "generate-release": "eslint-generate-release",
@@ -52,6 +63,7 @@
     "publish-release": "eslint-publish-release"
   },
   "repository": "eslint/eslint-visitor-keys",
+  "funding": "https://opencollective.com/eslint",
   "keywords": [],
   "author": "Toru Nagashima (https://github.com/mysticatea)",
   "license": "Apache-2.0",
